Refactor completed, added pom dependencies
