+++ project2/business/addbusinesscard/src/main/java/com/revature/BusinessCardMicroservice.java
@@ -1,0 +1,10 @@
+package com.revature;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BusinessCardMicroservice {
+
+    public static void main(String[] args) { SpringApplication.run(BusinessCardMicroservice.class, args); }
+}
