+++ project2/shared-libraries/dao/src/main/java/com/revature/dao/BusinessCardDAO.java
@@ -1,0 +1,10 @@
+package com.revature.dao;
+
+import com.revature.model.BusinessCard;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface BusinessCardDAO extends JpaRepository<BusinessCard, Integer> {
+
+}
