+++ project2/shared-libraries/dao/src/main/java/com/revature/dao/PersonalCardDAO.java
@@ -1,0 +1,7 @@
+package com.revature.dao;
+
+import com.revature.model.BusinessCard;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PersonalCardDAO extends JpaRepository<BusinessCard, Integer> {
+}
