--- project2/shared-libraries/daos/src/main/java/com/revature/daos/BusinessCardDAO.java
@@ -1,10 +1,0 @@
-package com.revature.daos;
-
-import com.revature.model.BusinessCard;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface BusinessCardDAO extends JpaRepository<BusinessCard, Integer> {
-
-}
