--- project2/shared-libraries/daos/src/main/java/com/revature/daos/PersonalCardDAO.java
@@ -1,7 +1,0 @@
-package com.revature.daos;
-
-import com.revature.model.BusinessCard;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PersonalCardDAO extends JpaRepository<BusinessCard, Integer> {
-}
