+++ project2/shared-libraries/utils/src/main/java/com/revature/util/JwtTokenFilter.java
@@ -1,0 +1,90 @@
+package com.revature.util;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import com.revature.model.User;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+import org.springframework.util.ObjectUtils;
+import org.springframework.web.filter.OncePerRequestFilter;
+
+import java.io.IOException;
+
+//This Util Class will make sure the requests hitting our server have a JWT in the Authorization header
+
+@Component //OncePerRequestFilter guarantees a single execution per request.
+public class JwtTokenFilter extends OncePerRequestFilter {
+    @Autowired //FIELD INJECTION BAD!
+    private JwtTokenUtil jwtUtil;
+
+    @Override
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response,
+                                    FilterChain filterChain) throws ServletException, IOException {
+
+        //If the Authorization header of the request doesn’t contain a Bearer token,
+        //it continues the filter chain without updating authentication context.
+        if (!hasAuthorizationBearer(request)) {
+            filterChain.doFilter(request, response);
+            return;
+        }
+
+        String token = getAccessToken(request);
+
+        //if the token is not verified,
+        //continue the filter chain without updating authentication context.
+        if (!jwtUtil.validateAccessToken(token)) {
+            filterChain.doFilter(request, response);
+            return;
+        }
+
+        //if the token is verified,
+        //update the authentication context with the user details ID and email.
+        setAuthenticationContext(token, request);
+        filterChain.doFilter(request, response);
+    }
+
+    //This is what extracts the Authorization header in our HTTP request to check for the JWT
+    private boolean hasAuthorizationBearer(HttpServletRequest request) {
+        String header = request.getHeader("Authorization");
+        if (ObjectUtils.isEmpty(header) || !header.startsWith("Bearer")) {
+            return false;
+        }
+        return true;
+    }
+
+    private String getAccessToken(HttpServletRequest request) {
+        String header = request.getHeader("Authorization");
+        String token = header.split(" ")[1].trim();
+        return token;
+    }
+
+    private void setAuthenticationContext(String token, HttpServletRequest request) {
+        UserDetails userDetails = getUserDetails(token);
+
+        UsernamePasswordAuthenticationToken
+                authentication = new UsernamePasswordAuthenticationToken(userDetails, null, null);
+
+        authentication.setDetails(
+                new WebAuthenticationDetailsSource().buildDetails(request));
+
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+
+    private UserDetails getUserDetails(String token) {
+        User userDetails = new User();
+        String[] jwtSubject = jwtUtil.getSubject(token).split(",");
+
+        userDetails.setUserId(Integer.parseInt(jwtSubject[0]));
+        userDetails.setUsername(jwtSubject[1]);
+
+        return userDetails;
+    }
+
+}
