+++ project2/shared-libraries/utils/src/main/java/com/revature/util/JwtTokenUtil.java
@@ -1,0 +1,64 @@
+package com.revature.util;
+
+import com.revature.model.User;
+import io.jsonwebtoken.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.util.Date;
+
+//This Util Class is what generates our JWTs
+@Component
+public class JwtTokenUtil {
+    private static final long EXPIRE_DURATION = 24 * 60 * 60 * 1000; // 24 hour
+
+    @Value("${app.jwt.secret}") //taken out of application.properties
+    private String SECRET_KEY;
+
+    private static final Logger LOGGER = LoggerFactory.getLogger(JwtTokenUtil.class);
+
+    //verify a given JWT. It returns true if the JWT is verified, or false otherwise.
+    public boolean validateAccessToken(String token) {
+        try {
+            Jwts.parser().setSigningKey(SECRET_KEY).parseClaimsJws(token);
+            return true;
+        } catch (ExpiredJwtException ex) {
+            LOGGER.error("JWT expired", ex.getMessage());
+        } catch (IllegalArgumentException ex) {
+            LOGGER.error("Token is null, empty or only whitespace", ex.getMessage());
+        } catch (MalformedJwtException ex) {
+            LOGGER.error("JWT is invalid", ex);
+        } catch (UnsupportedJwtException ex) {
+            LOGGER.error("JWT is not supported", ex);
+        } catch (SignatureException ex) {
+            LOGGER.error("Signature validation failed");
+        }
+
+        return false;
+    }
+
+    //Gets the value of the subject field of a given token.
+    //The subject contains User ID and email, which will be used to recreate a User object.
+    public String getSubject(String token) {
+        return parseClaims(token).getSubject();
+    }
+
+    private Claims parseClaims(String token) {
+        return Jwts.parser()
+                .setSigningKey(SECRET_KEY)
+                .parseClaimsJws(token)
+                .getBody();
+    }
+
+    public String generateAccessToken(User user) {
+        return Jwts.builder()
+                .setSubject(String.format("%s,%s", user.getUserId(), user.getUsername()))
+                .setIssuer("RevPay")
+                .setIssuedAt(new Date())
+                .setExpiration(new Date(System.currentTimeMillis() + EXPIRE_DURATION))
+                .signWith(SignatureAlgorithm.HS512, SECRET_KEY)
+                .compact();
+    }
+}
