+++ project2/shared-libraries/utils/src/main/java/com/revature/util/WebSecurityConfig.java
@@ -1,0 +1,80 @@
+package com.revature.util;
+
+//import com.revature.daos.UserDAO;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
+
+
+import javax.servlet.http.HttpServletResponse;
+
+
+/*
+This class extends the WebSecurityConfigurerAdapter.
+It's a convenience class that allows customization to both WebSecurity and HttpSecurity.
+
+This Util Class will define a couple things like password encryption and which routes need JWTs to access
+ */
+
+@EnableWebSecurity
+public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
+
+//    @Autowired
+//    private UserDAO userDAO;
+
+    @Autowired
+    private JwtTokenFilter jwtTokenFilter;
+
+//    @Override
+//    protected void configure(AuthenticationManagerBuilder auth) throws Exception {
+//        auth.userDetailsService(
+//                username -> (UserDetails) userDAO.findByUsername(username)
+//                        .orElseThrow(
+//                                () -> new UsernameNotFoundException("User " + username + " not found.")));
+//    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Override
+    @Bean
+    public AuthenticationManager authenticationManagerBean() throws Exception {
+        return super.authenticationManagerBean();
+    }
+
+    //This method configures which RESTful endpoints are universally accessible (login & add user)
+    //VS which endpoints need the user to be logged in for. (everything else)
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.csrf().disable();
+        http.sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS);
+
+        //here's the actual code that allows/restricts access to different endpoints/controllers
+        http.authorizeRequests()
+                .antMatchers("/auth/*").permitAll()
+//                .antMatchers("/admin/*").permitAll()
+                .anyRequest().authenticated();
+
+
+        http.exceptionHandling()
+                .authenticationEntryPoint(
+                        (request, response, ex) -> {
+                            response.sendError(
+                                    HttpServletResponse.SC_UNAUTHORIZED,
+                                    ex.getMessage()
+                            );
+                        }
+                );
+
+        http.addFilterBefore(jwtTokenFilter, UsernamePasswordAuthenticationFilter.class);
+    }
+}
