--- project2/src/main/java/com/revature/account/AccountMicroservice.java
@@ -1,9 +1,0 @@
-package com.revature.account;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AccountMicroservice {
-    public static void main(String[] args) { SpringApplication.run(AccountMicroservice.class, args); }
-}
