--- project2/src/main/java/com/revature/businessCard/BusinessCardMicroservice.java
@@ -1,10 +1,0 @@
-package com.revature.businessCard;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class BusinessCardMicroservice {
-
-    public static void main(String[] args) { SpringApplication.run(BusinessCardMicroservice.class, args); }
-}
