--- project2/src/main/java/com/revature/moneyRequest/MoneyRequestMicroservice.java
@@ -1,9 +1,0 @@
-package com.revature.moneyRequest;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MoneyRequestMicroservice {
-    public static void main(String[] args) { SpringApplication.run(MoneyRequestMicroservice.class, args); }
-}
