--- project2/src/main/java/com/revature/personalCard/PersonalCardMicroservice.java
@@ -1,10 +1,0 @@
-package com.revature.personalCard;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PersonalCardMicroservice {
-    public static void main(String[] args) { SpringApplication.run(PersonalCardMicroservice.class, args); }
-
-}
