--- project2/src/main/java/com/revature/transaction/TransactionMicroservice.java
@@ -1,9 +1,0 @@
-package com.revature.transaction;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TransactionMicroservice {
-    public static void main(String[] args) { SpringApplication.run(TransactionMicroservice.class, args); }
-}
