--- project2/src/main/java/com/revature/user/UserMicroservice.java
@@ -1,9 +1,0 @@
-package com.revature.user;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class UserMicroservice {
-    public static void main(String[] args) { SpringApplication.run(UserMicroservice.class, args); }
-}
