--- project2/src/main/java/com/revature/user/controller/UserController.java
@@ -1,4 +1,0 @@
-package com.revature.user.controller;
-
-public class UserController {
-}
