--- project2/src/main/resources/application.properties
@@ -1,34 +1,0 @@
-
-#Configuring the port this application will run on, and the base URL
-server.port = 8081
-server.servlet.context-path = /p2
-
-#Spring Boot Actuator Settings------------------
-
-#This lets us monitor the state of the application
-management.endpoint.health.show-details = always
-
-#This line lets us actually see the Actuator information
-management.endpoints.web.expose.include = *
-
-#Database Credentials------------------------------
-
-#These are the same credentials we're used to from JDBC
-spring.datasource.url = jdbc:postgresql://revpaydb.ccn4nz832krl.us-east-2.rds.amazonaws.com:5432/postgres
-spring.datasource.username = postgres
-spring.datasource.password = password
-
-#Spring Data JPA Settings-------------------------- SAVE FOR WHEN WE TALK ABOUT SPRING DATA
-
-#This will show the SQL running in our console. great for debugs
-spring.jpa.show-sql=true
-
-#Setting our DDL to update when a change happens in our model packages.
-#We could have set this to "create" which would erase your tables and recreate them when something changes
-spring.jpa.hibernate.ddl-auto=create
-
-#specify what schema we're pointing to This can change once the project starts
-spring.jpa.properties.hibernate.default_schema=public
-
-
-spring.jackson.serialization.FAIL_ON_EMPTY_BEANS=false
