--- project2/src/test/java/com/revature/project2/Project2ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.revature.project2;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Project2ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
