+++ project2/user/addpersonalcard/src/main/java/com/revature/PersonalCardMicroservice.java
@@ -1,0 +1,10 @@
+package com.revature;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PersonalCardMicroservice {
+    public static void main(String[] args) { SpringApplication.run(PersonalCardMicroservice.class, args); }
+
+}
